made auto-align more consistent?
I'm not really sure what the cause for this was; sometimes when auto-aligning to the step of switching between camera modes, the macro setting opens but doesn't navigate to the next tab. Could only be triggered on the first auto-align sequence

Just made the wait time a tiny bit longer, but it can also just be msedgeWebView2 causing the problem.